--- dbaClefPass.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Dec 30 16:14:47 2019
-
-@author: juan.cruz2
-"""
-
-import tkinter
-from tkinter import *
-from tkinter import ttk
-from tkinter import simpledialog
-
-
-root = tkinter.Tk() # dialog needs a root window, or will create an "ugly" one for you
-root.withdraw() # hide the root window
-
-root.destroy() # clean up after yourself!
